+++ changes.md
@@ -55,3 +55,7 @@
 ### 2016-09-27 zerobuilder 1.421 released
 
 * beans: remove singleton list shortcut for lone getters (breaking)
+
+### 2016-09-27 zerobuilder 1.431 released
+
+* rename `collection()` to `emptyCollection()` (beans) (lone getters) (breaking)